Don't create database by default
The container no longer creates administrator credentials. It is also
able to create a database even if the administrator user name and
password haven't been specified.

The rationale for this change is it should be possible to use this
image as a basis for a custom one where you might want to create
the administrator user and/or database and potentially populate it
with data or sync data from another instance - all these things
done as an anonymous user (CouchDB gives you full access until an
administrator account has been set up).

The scripts initialising the database have been heavily re-factored
replacing the procedural approach with a bunch of functions with
self-explanatory names.